+++ packages/zambdas/src/shared/candid.ts
@@ -39,7 +39,7 @@ import { Sex } from 'candidhealth/api/resources/preEncounter/resources/common/ty
 import { Coverage as CandidPreEncounterCoverage } from 'candidhealth/api/resources/preEncounter/resources/coverages/resources/v1/types/Coverage';
 import { MutableCoverage } from 'candidhealth/api/resources/preEncounter/resources/coverages/resources/v1/types/MutableCoverage';
 import { Patient as CandidPreEncounterPatient } from 'candidhealth/api/resources/preEncounter/resources/patients/resources/v1/types/Patient';
-import { ServiceLineCreate } from 'candidhealth/api/resources/serviceLines/resources/v2';
+import { MeasurementUnitCode, ServiceLineCreate } from 'candidhealth/api/resources/serviceLines/resources/v2';
 import { Operation } from 'fast-json-patch';
 import {
   Appointment,
@@ -74,9 +74,9 @@ import {
   getSecret,
   INVALID_INPUT_ERROR,
   isTelemedAppointment,
-  mapMedicationToCandidMeasurement,
   mapOrderStatusToFhir,
   MEDICATION_ADMINISTRATION_IN_PERSON_RESOURCE_CODE,
+  MedicationUnitOptions,
   MISSING_PATIENT_COVERAGE_INFO_ERROR,
   PaymentVariant,
   Secrets,
@@ -1407,4 +1407,19 @@ async function getMedicationAdministrationsForEncounter(
   ).unbundle();
   if (!resources || resources.length === 0) return undefined;
   return resources as MedicationAdministration[];
+}
+
+export function mapMedicationToCandidMeasurement(units: MedicationUnitOptions): MeasurementUnitCode | undefined {
+  switch (units) {
+    case 'mg':
+      return MeasurementUnitCode.Milligram;
+    case 'ml':
+      return MeasurementUnitCode.Milliliters;
+    case 'unit':
+      return MeasurementUnitCode.Units;
+    case 'g':
+      return MeasurementUnitCode.Grams;
+    default:
+      return MeasurementUnitCode.InternationalUnit; // todo ??? i have unhandled 'cc' and 'application' cases
+  }
 }
